Extend documentation around Flush and cached nodes
It took me quite a while to understand what's going on here, and I
agree with the existing comments from rvagg that it's probably not
very optimal.  There's not a lot in the way of options for controlling
memory usage by this "cache"; and it's also deeply critical to the
actual operating of writes, so it sort of feels like a misnomer to call
it "cache" at all.  However, I'm going to keep my mission in this
commit to "document" rather than "change": if nothing else, I think
some documentation about how to actually persist changes is useful,
because it wasn't at all obvious to me that it takes muliple function
calls with a variety of different parameters, and the order of those
calls is consequential.